+++ protected/humhub/modules/comment/notifications/NewComment.php
@@ -38,12 +38,21 @@ class NewComment extends \humhub\modules\notification\components\BaseNotificatio
     {
         return new CommentNotificationCategory();
     }
-
+    
+    public function hasMentioning(User $user)
+    {
+        return \humhub\modules\notification\models\Notification::find()->where([
+                    'class' => \humhub\modules\user\notifications\Mentioned::className(),
+                    'user_id' => $user->id,
+                    'source_class' => $this->source->className(),
+                    'source_pk' => $this->source->getPrimaryKey()])->count() > 0;
+    }
+    
     /**
      * @inheritdoc
      */
     public function send(User $user)
-    {
+    {   
         // Check if there is also a mention notification, so skip this notification
         if (\humhub\modules\notification\models\Notification::find()->where([
                     'class' => \humhub\modules\user\notifications\Mentioned::className(),